+++ mycodo/databases/models/misc.py
@@ -48,6 +48,7 @@ class Misc(CRUDMixin, db.Model):
     net_test_timeout = db.Column(db.Integer, default=3)
     default_login_page = db.Column(db.String, default='password')
     hostname_override = db.Column(db.String, default='')
+    custom_css = db.Column(db.String, default='')
 
     # Measurement database
     db_name = 'influxdb'  # Default